+++ zsh/.zshrc.d/30-aliases.zsh
@@ -3,5 +3,6 @@ alias ncu="npx npm-check-updates -i"
 alias lg="lazygit"
 alias ld="lazydocker"
 alias tf="terraform"
-alias nvp='NVIM_APPNAME=nvim-personal nvim'
+alias nvp='nvim'
+alias nvim-lazy='NVIM_APPNAME=lazyvim nvim'
 # alias pi='bun "$NVM_DIR/versions/node/$(node --version)/lib/node_modules/@earendil-works/pi-coding-agent/dist/cli.js"'